+++ src/icm20948/i2c.rs
@@ -471,11 +471,19 @@ where
     /// Rate must be less than 1.125kHz since that is the maximum of the accelerometer. Rate is specified in Hz.
     /// If rate is > 1.125kHz then an InvalidInput will be returned.
     /// Since the divisor is specified as an integer, the exact rate may not be met if it would require a floating point divisor.
+    ///
+    /// Note that while the gyro is enabled the gyro ODR (output data rate) will determine the interrupt frequency.
+    /// If the accelerometer frequency is lower than the gyro frequency then it will not be updated at every interrupt.
     pub fn config_acc_rate(&mut self, rate: u16) -> Result<(), IcmError<E>> {
         if rate < 1_125 {
             let div = 1_125 / (rate) - 1;
 
             self.change_bank(REG_BANK_2)?;
+
+            self.databuf[0] = RegistersBank2::OdrAlignEn.get_addr(WRITE_REG);
+            self.databuf[1] = 0x01;
+
+            self.bus.write(self.addr, &self.databuf[0..2])?;
 
             self.databuf[0] = RegistersBank2::AccelSmplrtDiv1.get_addr(WRITE_REG);
             self.databuf[1] = div.to_be_bytes()[0];
@@ -494,14 +502,23 @@ where
 
     /// Configure the data rate for the gyro.
     ///
-    /// Rate must be less than 1.1kHz since that is the maximum of the gyro. Rate is specified in Hz.
-    /// If rate is > 1.1kHz then an InvalidInput will be returned.
+    /// Rate must be less than 1.125kHz since that is the maximum of the gyro. Rate is specified in Hz.
+    /// If rate is > 1.125kHz then an InvalidInput will be returned.
+    /// Rate must be > 4 otherwise the divisor will not fit in a u8.
+    ///
     /// Since the divisor is specified as an integer, the exact rate may not be met if it would require a floating point divisor.
+    ///
+    /// Note that while the gyro is enabled the gyro ODR (output data rate) will determine the interrupt frequency.
     pub fn config_gyro_rate(&mut self, rate: u16) -> Result<(), IcmError<E>> {
-        if rate > 4 {
+        if rate > 4 && rate < 1125 {
             let div: u8 = (1_125 / (rate) - 1) as u8;
 
             self.change_bank(REG_BANK_2)?;
+
+            self.databuf[0] = RegistersBank2::OdrAlignEn.get_addr(WRITE_REG);
+            self.databuf[1] = 0x01;
+
+            self.bus.write(self.addr, &self.databuf[0..2])?;
 
             self.databuf[0] = RegistersBank2::GyroSmplrtDiv.get_addr(WRITE_REG);
             self.databuf[1] = div;
@@ -526,6 +543,11 @@ where
             div = div - 1;
 
             self.change_bank(REG_BANK_2)?;
+
+            self.databuf[0] = RegistersBank2::OdrAlignEn.get_addr(WRITE_REG);
+            self.databuf[1] = 0x01;
+
+            self.bus.write(self.addr, &self.databuf[0..2])?;
 
             self.databuf[0] = RegistersBank2::AccelSmplrtDiv1.get_addr(WRITE_REG);
             self.databuf[1] = div.to_be_bytes()[0];
@@ -552,6 +574,11 @@ where
 
         self.change_bank(REG_BANK_2)?;
 
+        self.databuf[0] = RegistersBank2::OdrAlignEn.get_addr(WRITE_REG);
+        self.databuf[1] = 0x01;
+
+        self.bus.write(self.addr, &self.databuf[0..2])?;
+
         self.databuf[0] = RegistersBank2::GyroSmplrtDiv.get_addr(WRITE_REG);
         self.databuf[1] = div;
 
@@ -565,12 +592,22 @@ where
     /// Resets the IMU the wakes it up from sleep mode.
     /// After the reset a 20ms sleep is suggested.
     ///
-    /// TODO: Unstable and NOT suggested for use.
+    /// If the IMU is not reset after writing the registers then the registers keep their same value.
+    /// This function is useful for testing code but currently should not be used in production.
+    /// You can also reset the IMU by power cycling it.
+    ///
+    /// TODO: NOT suggested for use.
     pub fn reset(&mut self) -> Result<(), IcmError<E>> {
         self.databuf[0] = RegistersBank0::PwrMgmt1.get_addr(WRITE_REG);
         self.databuf[1] = 0x80;
 
         self.bus.write(self.addr, &self.databuf[0..2])?;
+
+        // TODO: Replace with a delay
+        let mut j = 0;
+        for i in 1..2000000 {
+            j = j + i/10000;
+        }
 
         self.databuf[0] = RegistersBank0::PwrMgmt1.get_addr(WRITE_REG);
         self.databuf[1] = 0x01;