+++ changelog.md
@@ -8,8 +8,9 @@
 * More thorough testing on data rates and interrupts. The gyro data rate determines the update rate for the raw data interrupt if the gyro is enabled. 
 * Added interrupt examples.
 * Fixed enable and disable bug for accelerometer and gyroscope (for SPI).
-* Made the bus error more specific (TODO: Test this).
+* Made the bus error more specific.
 * Added an unstable reset function.
+* Improved documentation.
 
 ## v 0.1.0
 * Read accelerometer and gyroscope.